incident-communications-demo/backend/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
bidict==0.23.1
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
defusedxml==0.7.1
distro==1.9.0
Flask==3.0.0
Flask-Cors==5.0.0
Flask-SocketIO==5.3.6
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
MarkupSafe==3.0.2
msgpack==1.1.0
multidict==6.1.0
numpy==1.26.4
openai==1.61.0
orjson==3.10.15
packaging==23.2
presidio-analyzer==2.2.354
presidio-anonymizer==2.2.354
propcache==0.2.1
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.0
python-engineio==4.8.0
python-socketio==5.9.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
simple-websocket==1.1.0
sniffio==1.3.1
spacy==3.7.4
tenacity==8.5.0
tiktoken==0.8.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.18.3
zstandard==0.23.0
